Added fluid and NBT changes to the destruction gadget and planted infrastructure in to begin with the exchanging gadget
