+++ src/main/java/com/direwolf20/buildinggadgets/common/items/gadgets/GadgetDestruction.java
@@ -7,7 +7,6 @@ import com.direwolf20.buildinggadgets.common.config.Config;
 import com.direwolf20.buildinggadgets.common.entities.BlockBuildEntity;
 import com.direwolf20.buildinggadgets.common.registry.objects.BGBlocks;
 import com.direwolf20.buildinggadgets.common.tools.BlockMapIntState;
-import com.direwolf20.buildinggadgets.common.tools.LiquidInteractions;
 import com.direwolf20.buildinggadgets.common.utils.GadgetUtils;
 import com.direwolf20.buildinggadgets.common.utils.VectorUtil;
 import com.direwolf20.buildinggadgets.common.world.WorldSave;
@@ -24,7 +23,10 @@ import net.minecraft.nbt.NBTUtil;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.*;
 import net.minecraft.util.EnumFacing.Axis;
-import net.minecraft.util.math.*;
+import net.minecraft.util.math.AxisAlignedBB;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.RayTraceResult;
+import net.minecraft.util.math.Vec3i;
 import net.minecraft.util.text.ITextComponent;
 import net.minecraft.util.text.TextComponentString;
 import net.minecraft.util.text.TextComponentTranslation;
@@ -228,14 +230,6 @@ public class GadgetDestruction extends GadgetGeneric {
             }
         }
         return new ActionResult<ItemStack>(EnumActionResult.SUCCESS, stack);
-    }
-
-    public static RayTraceResult getLookingAt(EntityPlayer player, ItemStack stack) {
-        LiquidInteractions.GadgetGeneric FluidIntState = getFluidInteractionMode(stack);
-        if (FluidIntState == LiquidInteractions.GadgetGeneric.IGNORE) { //this is a binary system, so an if/else covers everything
-            return VectorUtil.getLookingAt(player);
-        }
-        return VectorUtil.getLookingAt(player, RayTraceFluidMode.ALWAYS);
     }
 
     public static void anchorBlocks(EntityPlayer player, ItemStack stack) {
